+++ src/MasterTool.Core/WishlistLogic.cs
@@ -1,0 +1,10 @@
+namespace MasterTool.Core;
+
+public static class WishlistLogic
+{
+    public static bool ShouldShowItem(bool wishlistOnly, bool isInWishlist, bool matchesTextFilter)
+    {
+        bool passesWishlist = !wishlistOnly || isInWishlist;
+        return passesWishlist && matchesTextFilter;
+    }
+}